`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: Migrate //chrome.
This CL makes use of the new `size_t`-based APIs in `//chrome`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5417393
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288014}

+++ chrome/browser/offline_pages/offline_page_url_loader.cc
@@ -9,6 +9,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/functional/bind.h"
 #include "base/memory/ptr_util.h"
 #include "base/notreached.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "chrome/browser/offline_pages/offline_page_utils.h"
 #include "chrome/browser/renderer_host/chrome_navigation_ui_data.h"
@@ -159,7 +160,7 @@ void OfflinePageURLLoader::NotifyReadRawDataComplete(int bytes_read) {
     return;
   }
 
-  bytes_of_raw_data_to_transfer_ = bytes_read;
+  bytes_of_raw_data_to_transfer_ = base::checked_cast<size_t>(bytes_read);
   write_position_ = 0;
 
   TransferRawData();
@@ -168,8 +169,7 @@ void OfflinePageURLLoader::NotifyReadRawDataComplete(int bytes_read) {
 void OfflinePageURLLoader::TransferRawData() {
   while (true) {
     DCHECK_GE(bytes_of_raw_data_to_transfer_, write_position_);
-    uint32_t write_size =
-        static_cast<uint32_t>(bytes_of_raw_data_to_transfer_ - write_position_);
+    size_t write_size = bytes_of_raw_data_to_transfer_ - write_position_;
     // If all the read data have been transferred, read more.
     if (write_size == 0) {
       ReadRawData();
